requirements.txt
dotmap==1.3.30
gif==3.0.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
matplotlib==3.5.1
numpy==1.21.5
omegaconf==2.2.2
pandas==1.4.1
plum_dispatch==1.5.9
pytest==7.1.2
PyYAML==6.0
scipy==1.7.3
setuptools==61.2.0
torch==1.13.0
torchdiffeq==0.2.3
tqdm==4.64.0
